Refactor country data import in Company and Individual components to use JSON file and map country names
