+++ src/ui/components/SingleAccount/company.tsx
@@ -24,9 +24,10 @@ import { PrimaryBtn, SecondaryBtn } from "../Buttons";
 import { DefaultAccount, validateAccount } from "@/lib/hooks/accounts";
 import DropzoneComponent from "../Dropzone";
 import { sendMoneyCompanyValidate } from "@/lib/schema";
-import { countries } from "@/lib/static";
+// import { countries } from "@/lib/static";
 import { useDebouncedValue } from "@mantine/hooks";
 import { removeWhitespace } from "@/lib/utils";
+import countries from "@/assets/countries.json";
 
 interface CompanyProps {
   account: DefaultAccount | null;
@@ -281,7 +282,7 @@ function Company({
                   classNames={{ input: styles.input, label: styles.label }}
                   flex={1}
                   label="Country"
-                  data={countries}
+                  data={countries.map((c) => c.name)}
                   disabled={disableCountry}
                   {...form2.getInputProps("destinationCountry")}
                 />
